requirements.txt
python==3.6.8
beautifulsoup4==4.11.1
dash==2.3.1
dash_core_components==2.0.0
dash_html_components==2.0.0
Flask==2.1.0
matplotlib==3.3.2
numpy
pandas
pandas_datareader==0.10.0
plotly==5.7.0
python-decouple==3.6
requests==2.24.0
scikit_learn==1.1.1
tensorflow-cpu
transformers==4.18.0
gunicorn == 20.0.4